+++ mysite02/src/main/java/com/douzone/mysite/mvc/board/DeleteAction.java
@@ -15,7 +15,7 @@ public class DeleteAction implements Action {
 	@Override
 	public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		
-		Long no = Long.parseLong(request.getParameter("no"));
+		int no = Integer.parseInt(request.getParameter("no"));
 		
 		new BoardDao().delete(no);
