+++ mysite02/src/main/java/com/douzone/mysite/vo/UserVo.java
@@ -1,16 +1,17 @@
 package com.douzone.mysite.vo;
 
 public class UserVo {
-	private Long no;
+	private int no;
 	private String name;
 	private String email;
 	private String password;
 	private String gender;
 	private String joinDate;
-	public Long getNo() {
+	
+	public int getNo() {
 		return no;
 	}
-	public void setNo(Long no) {
+	public void setNo(int no) {
 		this.no = no;
 	}
 	public String getName() {
